Minor updates in the animations classes of AnimationsUtility and ToolAnimationsUtility

+++ app/src/main/java/com/nerdcoredevelopment/game2048champsfinal/animations/AnimationsUtility.java
@@ -78,7 +78,7 @@ public class AnimationsUtility {
         setupGamePreviewAnimator.start();
     }
 
-    private static void setTextViewAttributes(AppCompatTextView textView, int cellValue,
+    private static void setTextViewAttributes(AppCompatTextView textView, long cellValue,
                                               int textColor, Drawable backgroundDrawable,
                                               GameLayoutProperties gameLayoutProperties) {
         textView.setVisibility(View.VISIBLE);
@@ -91,7 +91,7 @@ public class AnimationsUtility {
     /* Note: ViewGroup is the parent class for all the Layout classes and View is the parent class
              for all types of views
     */
-    public static void executePopUpAnimation(AppCompatTextView textView, int cellValue, int textColor,
+    public static void executePopUpAnimation(AppCompatTextView textView, long cellValue, int textColor,
                                              Drawable backgroundDrawable, long duration, long delay,
                                              GameLayoutProperties gameLayoutProperties) {
         ObjectAnimator popUpXAnimator = ObjectAnimator.ofFloat(textView, View.SCALE_X, 0.5f, 1f).setDuration(duration);
@@ -204,7 +204,7 @@ public class AnimationsUtility {
     /**
      * The following two are end part animations
      */
-    public static AnimatorSet getSimplyAppearAnimation(AppCompatTextView textView, int cellValue, int textColor,
+    public static AnimatorSet getSimplyAppearAnimation(AppCompatTextView textView, long cellValue, int textColor,
                                                        Drawable backgroundDrawable, long duration, long delay,
                                                        GameLayoutProperties gameLayoutProperties) {
         AnimatorSet simplyAppearAnimatorSet = new AnimatorSet();
@@ -226,7 +226,7 @@ public class AnimationsUtility {
         return simplyAppearAnimatorSet;
     }
 
-    public static AnimatorSet getMergeAnimation(AppCompatTextView textView, int cellValue, int textColor,
+    public static AnimatorSet getMergeAnimation(AppCompatTextView textView, long cellValue, int textColor,
                                                 Drawable backgroundDrawable, long duration, long delay,
                                                 GameLayoutProperties gameLayoutProperties) {
         AnimatorSet mergeAnimatorSet = new AnimatorSet();
